+++ HISTORY.md
@@ -1,5 +1,8 @@
 ### mlpack ?.?.?
 ###### ????-??-??
+  * Added Decision Tree Regressor (#2905). It can be used using the class
+    `mlpack::tree::DecisionTreeRegressor`. It is accessible only though C++.
+    
   * Added dict-style inspection of mlpack models in python bindings (#2868).
 
   * Added Extra Trees Algorithm (#2883). Currently, it can be used using the
@@ -60,6 +63,11 @@
     constructor is used (#2964).
 
   * Fix Julia model serialization bug (#2970).
+
+  * Fix `LoadCSV()` to use pre-populated `DatasetInfo` objects (#2980).
+
+  * Add `probabilities` option to softmax regression binding, to get class
+    probabilities for test points (#3001).
 
 ### mlpack 3.4.2
 ###### 2020-10-26